requirements.txt
setuptools==70.0.0
rich==10.0.0
dataclasses~=0.6
requests~=2.32.3
beautifulsoup4~=4.12.3
PyYAML~=6.0.2
